moved applicationForManager type
